requirements.txt
transformers
torch
Pillow
soundfile
numpy
pandas
tqdm